Fix bugs and do some clean
