+++ GlassCannon.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31424.327
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GlassCannon", "GlassCannon\GlassCannon.csproj", "{7DD3DB62-C91E-4737-A799-C4777B13A2B1}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{7DD3DB62-C91E-4737-A799-C4777B13A2B1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7DD3DB62-C91E-4737-A799-C4777B13A2B1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7DD3DB62-C91E-4737-A799-C4777B13A2B1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7DD3DB62-C91E-4737-A799-C4777B13A2B1}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {FA5B1A05-8C78-4504-98C8-500C0035236E}
+	EndGlobalSection
+EndGlobal
